Detangle papis.el from README.org
While I appreciate the literate programming effort of including the
whole 'papis.el' code in the README.org, I think it works against the
project by adding a barrier for contributions, and obfuscating the
actual elisp code with backlinks to the README. In this case, the
audience for the README mainly includes users who want to know how to
install and configure the library; the lisp hackers are used dig
through the .el files.

Tangling can still remain useful for the configuration instructions,
so I don't remove it altogether here.