'doc' accessors: Rename out of public API, docstring and test
(papis-doc-get): Rename to `papis--doc-get'
(papis-doc-get-folder): Rename to `papis--doc-folder'
(papis--get-ref): Rename to `papis--doc-ref'
(papis-doc-id): Rename to `papis--doc-id'
(papis-id-query): Rename to `papis--doc-query'

(papis--get-file-paths): Rename to `papis--doc-file-paths' and check
file existence.

(papis--notes-path): Rename to `papis--doc-notes-path' and avoid
creating a Papis process.

(papis--ensured-notes-path): Simplify and accept a more general
argument "query" so that it can be called with a citekey in the future
Papis/Citar integration.